only create `profile.d` layer if the directory exists
the cnb-shim will always try to copy the `.profile.d` directory over and
create a layer for it, but this will lead to an error if the buildpack
doesn't create it.

```
cp: cannot stat '.profile.d/*': No such file or directory
```

This change now conditionally creates the layer if the buildpack is
created a `.profile.d` directory.