Add logging statements to the shim
Right now then code in the shim fails it looks like the error is inside of the buildpack itself. For example this error:

```
[builder] -----> Detecting rails configuration
[builder] exit status 127
[builder] ERROR: failed to build: exit status 3
ERROR: failed to build: executing lifecycle. This may be the result of using an untrusted builder: failed with status code: 51
```

Appears to indicate that the problem happened in `bin/compile` when in fact that was successful and the real problem didn't happen until `bin/release`. By logging our intentions to the output it makes it clearer to understand where failures are happening if/when they do happen.