Ensure generated archives are deterministic
Previously each request to `cnb-shim` for the same buildpack would
result in a tarfile with varying contents even if the underlying buildpack
hadn't changed).

This was due to:
- the dynamically generated parts of the shimmed buildpack
  having "now" timestamps
- the archive containing the user ID of the current dyno, which
  varies between dynos (and requests to cnb-shim can hit any
  of its `web` dynos, plus dynos are cycled every 24 hours).

Now, repeat requests result in an identical archive, so long as the
buildpack or shim scripts haven't changed.

This helps reduce layer churn in builder images containing shimmed
buildpacks, such as our new `heroku/builder-classic:22` builder, or
the older `heroku/buildpacks:{18,22}` builders.

GUS-W-11277378.