Fix the default `name` if none is specified via a URL parameter
Previously if no `&name=...` parameter was specified in the URL, then
the default buildpack `name` (the string used for display purposes only,
not to be confused by the buildpack `id`) defaulted to the `"0.1"`, which
appears to be a copy-paste mistake when the URL handling was added.

Now, if no `name` is specified, it defaults to the same value as the buildpack
`id`, which will be less pretty in the build logs than a user-provided name
(such as "Python"), but at least clearer than "0.1".

GUS-W-12541186.